+++ lib/src/connectionhandling/server/mqtt_client_mqtt_server_ws_connection.dart
@@ -59,7 +59,7 @@ class MqttServerWsConnection extends MqttServerConnection<WebSocket> {
           .then((socket) {
         client = socket;
         readWrapper = ReadWrapper();
-        messageStream = MqttByteBuffer(typed.Uint8Buffer());
+        messageStream = MqttByteBuffer(typed.Uint32Buffer());
         _startListening();
         completer.complete();
       }).catchError((e) {
