+++ test/mqtt_client_connection_secure_test.dart
@@ -9,13 +9,15 @@
 library;
 
 import 'dart:io';
+
+import 'package:event_bus/event_bus.dart' as events;
+import 'package:mocktail/mocktail.dart';
 import 'package:mqtt_client/mqtt_client.dart';
 import 'package:mqtt_client/mqtt_server_client.dart';
-import 'package:test/test.dart';
-import 'package:mocktail/mocktail.dart';
-import 'package:typed_data/typed_data.dart' as typed;
 import 'package:path/path.dart' as path;
-import 'package:event_bus/event_bus.dart' as events;
+import 'package:test/test.dart';
+import 'package:typed_data/typed_data.dart' as typed;
+
 import 'support/mqtt_client_mockbroker.dart';
 
 // Mock classes
@@ -47,7 +49,7 @@ void main() {
     setUp(() async {
       broker = MockBrokerSecure();
       broker.pemName = 'localhost';
-      void messageHandlerConnect(typed.Uint8Buffer? messageArrived) {
+      void messageHandlerConnect(typed.Uint32Buffer? messageArrived) {
         final ack = MqttConnectAckMessage()
             .withReturnCode(MqttConnectReturnCode.connectionAccepted);
         broker.sendMessage(ack);
@@ -63,7 +65,7 @@ void main() {
     test('Connection Keep Alive - Successful response', () async {
       var expectRequest = 0;
 
-      void messageHandlerPingRequest(typed.Uint8Buffer? messageArrived) {
+      void messageHandlerPingRequest(typed.Uint32Buffer? messageArrived) {
         final headerStream = MqttByteBuffer(messageArrived);
         final header = MqttHeader.fromByteBuffer(headerStream);
         if (expectRequest <= 3) {
